requirements.txt
Django==2.0.7
django-cors-headers==2.3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
graphene-django==2.0.0
grpcio-tools==1.13.0
pylint-django==0.11.1
pytest-django==3.3.2
python-bitcoinrpc==1.0
yapf==0.22.0